+++ NEWS.md
@@ -1,4 +1,7 @@
 # noctua 1.10.999
+## New Feature:
+* Added support to `AWS Athena` data types array, row and map (#135).
+
 ## Bug Fix:
 * `AWS Athena` uses `float` data type for the DDL only, `noctua` was wrongly parsing `float` data type back to R. Instead `AWS Athena` uses data type `real` in SQL functions like `select cast` https://docs.aws.amazon.com/athena/latest/ug/data-types.html. `nocuta` now correctly parses `real` to R's data type `double` (#133)